+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [0.1.0.nullsafety.0]
 
 * Migrates to null-safety
+* Renames `Router -> PathRouter`
 
 ## [0.0.1+3]